Restructure postbox classes
